requirements.txt
Django==2.1
django-appconf==1.0.2
django-crispy-forms==1.7.2
django-suit==2.0a1
django-suit-redactor==0.0.4
mysqlclient==1.3.13
Pillow==5.1.0
pytz==2018.4
